requirements.txt
a2a-sdk[all]
langchain-databricks>=0.1.0
langchain-community>=0.2.0
langchain>=0.2.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
databricks-vectorsearch
databricks-sdk
mlflow>=2.10.0
python-dotenv
uvicorn
fastapi
pydantic_settings